Face Recognition/requirements.txt
#test deploy again
firebase_admin==6.2.0
numpy==1.24.3
onnxruntime==1.22.0
opencv-python==4.8.0.76
Pillow>=7.1.0,<10.0.0
scikit-learn==1.3.2
streamlit==1.25.0
streamlit-option-menu==0.3.6
streamlit-webrtc==0.45.0
torch==2.1.2
ultralytics==8.0.172
gunicorn==20.1.0
flask==2.3.3
